# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Base Eleitoral 360 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /zonas:
    get:
      operationId: listZonas
      tags: [zonas]
      summary: List electoral zones
      responses:
        "200":
          description: List of zones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ZonaEleitoral"
    post:
      operationId: createZona
      tags: [zonas]
      summary: Create an electoral zone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZonaEleitoralInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZonaEleitoral"
  /zonas/{id}:
    patch:
      operationId: updateZona
      tags: [zonas]
      summary: Update an electoral zone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZonaEleitoralUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZonaEleitoral"
    delete:
      operationId: deleteZona
      tags: [zonas]
      summary: Delete an electoral zone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Agenda de Campo ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ZonaEleitoral:
      type: object
      required: [id, zona, bairro, cidade, estado, totalEleitores]
      properties:
        id:
          type: integer
        zona:
          type: string
        secao:
          type: ["string", "null"]
        localVotacao:
          type: ["string", "null"]
        endereco:
          type: ["string", "null"]
        bairro:
          type: string
        cidade:
          type: string
        estado:
          type: string
        totalEleitores:
          type: integer
        historicoVotacao:
          type: ["string", "null"]
        metaVotos:
          type: ["integer", "null"]
        votosEstimados:
          type: ["integer", "null"]
        votosValidados:
          type: ["integer", "null"]
        responsavel:
          type: ["string", "null"]
        observacoes:
          type: ["string", "null"]
    ZonaEleitoralInput:
      type: object
      required: [zona, bairro, cidade, estado, totalEleitores]
      properties:
        zona:
          type: string
        secao:
          type: string
        localVotacao:
          type: string
        endereco:
          type: string
        bairro:
          type: string
        cidade:
          type: string
        estado:
          type: string
        totalEleitores:
          type: integer
        historicoVotacao:
          type: string
        metaVotos:
          type: integer
        votosEstimados:
          type: integer
        votosValidados:
          type: integer
        responsavel:
          type: string
        observacoes:
          type: string
    ZonaEleitoralUpdate:
      type: object
      properties:
        zona:
          type: string
        secao:
          type: string
        localVotacao:
          type: string
        endereco:
          type: string
        bairro:
          type: string
        cidade:
          type: string
        estado:
          type: string
        totalEleitores:
          type: integer
        historicoVotacao:
          type: string
        metaVotos:
          type: integer
        votosEstimados:
          type: integer
        votosValidados:
          type: integer
        responsavel:
          type: string
        observacoes:
          type: string

    # ─── Agenda ────────────────────────────────────────────────────────────────
...