# lib/api-zod/src/generated/types/zonaEleitoralInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Base Eleitoral 360 API
 * OpenAPI spec version: 0.1.0
 */

export interface ZonaEleitoralInput {
  zona: string;
  secao?: string;
  localVotacao?: string;
  endereco?: string;
  bairro: string;
  cidade: string;
  estado: string;
  totalEleitores: number;
  historicoVotacao?: string;
  metaVotos?: number;
  votosEstimados?: number;
  votosValidados?: number;
  responsavel?: string;
  observacoes?: string;
}